fix: #305 중복 transactional 오류
